requirements.txt
fastapi
uvicorn
ultralytics
pillow
cloudinary
python-dotenv
